+++ src/TestUtils/AuthorizationTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\AuthorizationBundle\TestUtils;
 
-use Dbp\Relay\AuthorizationBundle\Authorization\AuthorizationService;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 class AuthorizationTest
@@ -12,25 +11,5 @@ class AuthorizationTest
     public static function setUp(ContainerInterface $container): void
     {
         TestEntityManager::setUpAuthorizationEntityManager($container);
-    }
-
-    public static function tearDown(ContainerInterface $container): void
-    {
-        $authorizationService = self::getAuthorizationService($container);
-        $authorizationService->clearRequestCache();
-        $authorizationService->getCache()->clear();
-    }
-
-    public static function postRequestCleanup(ContainerInterface $container): void
-    {
-        self::getAuthorizationService($container)->clearRequestCache();
-    }
-
-    private static function getAuthorizationService(ContainerInterface $container): AuthorizationService
-    {
-        $authorizationService = $container->get(AuthorizationService::class);
-        assert($authorizationService instanceof AuthorizationService);
-
-        return $authorizationService;
     }
 }